weird git file rename bug resolution
